Support the latest Firebase SDK versions.
There are two compromises:

1. I gave up on monkeypatching the SDK and instead went with an explicit
function call to wrap a Database object.  The APIs of the client- and
server-side Firebase SDK have diverged, and the team has stated that it's
their attention to diverge further.  Monkeypatching becomes too brittle
in that scenario, so it's now on library clients to ensure that only a
wrapped instance is used.

2. I still use the "compatibility" variant of the client-side API.  This
will be deprecated and possibly removed in the future, at which point
we'll need to fork this library into server and client variants, but
let's cross that bridge when it's thrust upon us (to mix metaphors).