Fix isEqual comparisons between queries and references.
The Firebase API allows isEqual comparisons between queries and references,
though they'll always return false.  On the other hand, it appears that it's
very unhappy if you pass in `undefined` (though `null` is OK).  So let's
extract either the underlying query or reference from the argument that got
passed in, maintaining the original's behavior as much as possible.