Decrypt and decompress values that are no longer flagged in the spec.
We need to keep recursing even after running out of spec defs, since
the `encrypted` attribute might've been removed from a value that was
previously encrypted or compressed.  The downside is that this will
recreate all objects when transforming (even ones that don't need
any decryption) but the only alternative is to forbid removing
`encrypted` attributes from the spec altogether, which seems
draconian.